+++ js/main.js
@@ -46,38 +46,9 @@ const onGlobalKeyDown = (evt) => {
 document.addEventListener('click', onGlobalClick);
 document.addEventListener('keydown', onGlobalKeyDown);
 
-const showError = (message) => {
-  const errorContainer = document.createElement('div');
-  errorContainer.style.zIndex = 100;
-  errorContainer.style.position = 'absolute';
-  errorContainer.style.left = 0;
-  errorContainer.style.top = 0;
-  errorContainer.style.right = 0;
-  errorContainer.style.padding = '10px 3px';
-  errorContainer.style.fontSize = '30px';
-  errorContainer.style.textAlign = 'center';
-  errorContainer.style.backgroundColor = 'red';
 
-  errorContainer.textContent = message;
-
-  document.body.append(errorContainer);
-
-  // setTimeout(() => {
-  //   alertContainer.remove();
-  // }, ALERT_SHOW_TIME);
-};
-
-
-// const error = () => {
-
-// };
-
-const loadPhoto = getData((renderPhotoList),showError('Ошибка загрузки данных. Попробуйте позже'));
-loadPhoto();
-
-// getData((cardPhoto) => {
-//   renderPhotoList(cardPhoto);
-//   showError('Ошибка загрузки данных. Попробуйте позже');
-// });
+getData((cardPhoto) => {
+  renderPhotoList(cardPhoto);
+});
 
 setUserFormSubmit(closeForm);
