+++ js/util.js
@@ -26,11 +26,28 @@ const toggleVisibleBigPicture = (flag, className) => {
 
 };
 
+const showError = (message) => {
+  const errorContainer = document.createElement('div');
+  errorContainer.style.zIndex = 100;
+  errorContainer.style.position = 'absolute';
+  errorContainer.style.left = 0;
+  errorContainer.style.top = 0;
+  errorContainer.style.right = 0;
+  errorContainer.style.padding = '10px 3px';
+  errorContainer.style.fontSize = '30px';
+  errorContainer.style.textAlign = 'center';
+  errorContainer.style.backgroundColor = 'red';
+
+  errorContainer.textContent = message;
+
+  document.body.append(errorContainer);
+};
 
 export {
   getRandomInteger,
   getRandomArrayElement,
   maxLengthString,
   getTemplate,
-  toggleVisibleBigPicture
+  toggleVisibleBigPicture,
+  showError
 };
